refactor: moved the buttons and cards into the components folder and updated their imports
