+++ client/src/components/cards/UserCard.tsx
@@ -1,7 +1,7 @@
 import { useState, useEffect, useContext } from 'react';
 import { Col, Row } from 'react-bootstrap';
 import axios from 'axios';
-import { UserContext } from '../context';
+import { UserContext } from '../../context';
 import FollowButton from '../buttons/FollowButton';
 import BlockButton from '../buttons/BlockButton';
 import DmButton from '../buttons/DmButton';
@@ -13,7 +13,7 @@ import {
   StyledPartyTitle,
   StyledUserCardFooter,
   StyledUserCardFooterCol,
-} from '../cards/cards.styles';
+} from './cards.styles';
 
 // ModCard stands for Modular Card. Hopefully we can reuse it
 function UserCard({ obj, socket }) {
