+++ client/src/views/search/Search.tsx
@@ -1,8 +1,8 @@
 import { useContext } from 'react';
 import { Container, Col, Row } from 'react-bootstrap';
-import UserCard from '../../cards/UserCard';
-import PartyCard from '../../cards/PartyCard';
-import VideoCard from '../../cards/VideoCard';
+import UserCard from '../../components/cards/UserCard';
+import PartyCard from '../../components/cards/PartyCard';
+import VideoCard from '../../components/cards/VideoCard';
 import { SearchContext } from '../../contexts/searchContext';
 import { UserContext } from '../../context';
 import { CategoryTitle } from '../../styles';
